Advance-Claude-Development-phase-5-main/backend/requirements.txt
fastapi[standard]>=0.115.0
sqlmodel>=0.0.22
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
uvicorn[standard]>=0.32.0
alembic>=1.14.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
python-jose[cryptography]>=3.3.0
openai-agents>=0.1.0
structlog>=24.4.0
